+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/workrecord/erroralarm/condition/attendanceitem/KrcmtErAlAtdItemConPK.java
@@ -4,7 +4,6 @@
 package nts.uk.ctx.at.record.infra.entity.workrecord.erroralarm.condition.attendanceitem;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 import javax.persistence.Basic;
 import javax.persistence.Column;
@@ -33,6 +32,6 @@ public class KrcmtErAlAtdItemConPK implements Serializable {
 	@Basic(optional = false)
 	@NotNull
 	@Column(name ="ATD_ITEM_CON_NO")
-	public BigDecimal atdItemConNo;
+	public int atdItemConNo;
 	
 }
